Refactoring error handling for IPC and services
