+++ binder/src/ipc/app.rs
@@ -24,16 +24,16 @@ use sgx_types::{sgx_enclave_id_t, sgx_status_t};
 use crate::ipc::IpcError;
 use crate::ipc::IpcSender;
 use log::{debug, error};
-use teaclave_types::EnclaveStatus;
+use teaclave_types::ECallStatus;
 
 // Delaration of ecall for App, the implementation is in TEE
 // This function is automatically generated by the procedure macro #[ecall_entry_point].
-// fn enclave_init(eid: sgx_enclave_id_t, retval: *mut EnclaveStatus) -> sgx_status_t;
+// fn enclave_init(eid: sgx_enclave_id_t, retval: *mut ECallStatus) -> sgx_status_t;
 #[no_mangle]
 extern "C" {
     fn ecall_ipc_entry_point(
         eid: sgx_enclave_id_t,
-        retval: *mut EnclaveStatus,
+        retval: *mut ECallStatus,
         cmd: u32,
         in_buf: *const u8,
         in_len: usize,
@@ -75,7 +75,7 @@ impl ECallChannel {
             let mut out_len: usize = out_max;
             let out_ptr: *mut u8 = out_buf.as_mut_ptr();
 
-            let mut ecall_ret = EnclaveStatus::default();
+            let mut ecall_ret = ECallStatus::default();
 
             let sgx_status = unsafe {
                 ecall_ipc_entry_point(
@@ -93,7 +93,7 @@ impl ECallChannel {
             // Check sgx return values
             if sgx_status != sgx_status_t::SGX_SUCCESS {
                 error!("ecall_ipc_entry_point, app sgx_error:{}", sgx_status);
-                return Err(IpcError::SgxError(sgx_status as i32));
+                return Err(IpcError::SgxError(sgx_status));
             }
 
             // Check rust logic return values
@@ -115,7 +115,7 @@ impl ECallChannel {
             // Transparent deliever the errors to outer logic.
             if ecall_ret.is_err() {
                 error!("ecall_ipc_entry_point, app api_error: {:?}", ecall_ret);
-                return Err(IpcError::EnclaveError(ecall_ret));
+                return Err(IpcError::ECallError(ecall_ret));
             }
 
             unsafe {
@@ -136,9 +136,9 @@ impl IpcSender for ECallChannel {
         U: Serialize,
         V: for<'de> Deserialize<'de>,
     {
-        let request_payload = serde_json::to_vec(&input).map_err(|_| IpcError::SerdeError)?;
+        let request_payload = serde_json::to_vec(&input)?;
         let result_buf = self.ecall_ipc_app_to_tee(cmd, request_payload)?;
-        let response: V = serde_json::from_slice(&result_buf).map_err(|_| IpcError::SerdeError)?;
+        let response: V = serde_json::from_slice(&result_buf)?;
         Ok(response)
     }
 }
