+++ binder/src/ipc/enclave.rs
@@ -31,8 +31,7 @@ impl IpcReceiver for ECallReceiver {
         X: IpcService<U, V>,
     {
         let input: U = serde_json::from_slice(&input_payload)?;
-        let response: std::result::Result<V, teaclave_types::TeeServiceError> =
-            x.handle_invoke(input);
+        let response: Result<V, teaclave_types::TeeServiceError> = x.handle_invoke(input);
         let response_payload = serde_json::to_vec(&response)?;
 
         Ok(response_payload)
