+++ services/management/enclave/src/lib.rs
@@ -42,6 +42,7 @@ use teaclave_rpc::server::SgxTrustedTlsServer;
 use teaclave_service_enclave_utils::{create_trusted_storage_endpoint, ServiceEnclave};
 use teaclave_types::{EnclaveInfo, TeeServiceError, TeeServiceResult};
 
+mod error;
 mod service;
 
 fn start_service(config: &RuntimeConfig) -> Result<()> {
