+++ services/storage/enclave/src/lib.rs
@@ -45,6 +45,7 @@ use teaclave_rpc::server::SgxTrustedTlsServer;
 use teaclave_service_enclave_utils::ServiceEnclave;
 use teaclave_types::{EnclaveInfo, TeeServiceError, TeeServiceResult};
 
+mod error;
 mod proxy;
 mod service;
